Turn off checking of_node for irq interrupts
Currently do not have interrupts defined in hardware and thus they are
not defined in the devicetree. Since the code/kernel thinks there should
be interrupts defined in the devicetree node it throws an error when
there is not. There should be a better way to do this but this works for
now.