Add script to automate simple IP Package creation
Invoking this script with a modules filename as argument will try and create a
Xilinx IP Package. Once crated the package is moved to a 'ip/ip_repo' directory
and can be used in a Vivado project.

The method of package creation is simple and for the most part automatic and
thus we assume that Xilinx naming conventions etc are observed.