Bug fix: processing consecutive streams of read data
The seriliser needs to be reset when there is leftover data, however, if the
module is processing multiple streams of data that have been read from memory
there might be valid data that has been popped from the data fifo. If that is
the case we must not reset the seriliser.