Bug fix: axis_write_data contiguous streams
When a stream comes from the user system and is done there is a possibility
that the first data word of the next stream also gets popped from the buffer.
In those cases we need to take note of this so that the 'str_cnt' can be
correctly incremented and the deser_(done|last) values correctly reset.